pybind11: refactor variant caster
