+++ include/openPMD/binding/python/Numpy.hpp
@@ -20,11 +20,12 @@
  */
 #pragma once
 
+#include "openPMD/Datatype.hpp"
+#include "openPMD/binding/python/Variant.hpp"
+
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 #include <pybind11/numpy.h>
-
-#include "openPMD/Datatype.hpp"
 
 #include <string>
 #include <exception>
